+++ packages/svelte/src/main/main-server.js
@@ -20,3 +20,13 @@ export function beforeUpdate() {}
 
 /** @returns {void} */
 export function afterUpdate() {}
+
+/**
+ * @template T
+ * @param {T} value
+ * @returns {T}
+ */
+export function unstate(value) {
+	// There's no signals/proxies on the server, so just return the value
+	return value;
+}