+++ packages/svelte/tests/runtime-runes/samples/snippet-store/main.svelte
@@ -1,0 +1,11 @@
+<script>
+	import { writable } from 'svelte/store';
+
+	let snippet = writable(hello);
+</script>
+
+{#snippet hello()}
+	<p>hello world</p>
+{/snippet}
+
+{@render $snippet()}